Issue 54: FT HTTP - delay for user acceptance should depend on file transfer validity Fixes issue 54 - Save file and icon expiration timestamp into filetransfer provider. - Wait for user acceptance with a timeout depending on the expiration time. - Upon user acceptance, retrieve session from provider if it is not ongoing. - Create new session objects to start a HTTP terminating File transfer session from INVITE, from ACCEPT or from RESUME. - Save remote SIP Id in provider for later delivery notification through SIP message. - Save mime type for file icon (bug fix). - Modify selection in GroupDeliveryInfoLog (bug fix).

+++ core/src/com/gsma/rcs/provider/messaging/GroupDeliveryInfoLog.java
@@ -42,15 +42,17 @@ public class GroupDeliveryInfoLog implements IGroupDeliveryInfoLog {
             .append(GroupDeliveryInfoData.KEY_CONTACT).append("=?").toString();
 
     private static final String SELECTION_CONTACTS_NOT_RECEIVED_MESSAGE = new StringBuilder(
-            GroupDeliveryInfoData.KEY_DELIVERY_STATUS).append("=").append(Status.NOT_DELIVERED)
-            .append(" OR (").append(GroupDeliveryInfoData.KEY_DELIVERY_STATUS).append("=")
-            .append(Status.FAILED).append(" AND ").append(GroupDeliveryInfoData.KEY_REASON_CODE)
-            .append(" IN (").append(ReasonCode.FAILED_DELIVERY).append(",")
-            .append(ReasonCode.FAILED_DISPLAY).append("))").toString();
+            GroupDeliveryInfoData.KEY_DELIVERY_STATUS).append("=")
+            .append(Status.NOT_DELIVERED.toInt()).append(" OR (")
+            .append(GroupDeliveryInfoData.KEY_DELIVERY_STATUS).append("=")
+            .append(Status.FAILED.toInt()).append(" AND ")
+            .append(GroupDeliveryInfoData.KEY_REASON_CODE).append(" IN (")
+            .append(ReasonCode.FAILED_DELIVERY.toInt()).append(",")
+            .append(ReasonCode.FAILED_DISPLAY.toInt()).append("))").toString();
 
     private static final String SELECTION_DELIVERY_INFO_NOT_DISPLAYED = new StringBuilder(
-            GroupDeliveryInfoData.KEY_DELIVERY_STATUS).append("!=").append(Status.DISPLAYED)
-            .toString();
+            GroupDeliveryInfoData.KEY_DELIVERY_STATUS).append("!=")
+            .append(Status.DISPLAYED.toInt()).toString();
 
     private static final Logger logger = Logger.getLogger(GroupDeliveryInfoLog.class
             .getSimpleName());
@@ -83,7 +85,7 @@ public class GroupDeliveryInfoLog implements IGroupDeliveryInfoLog {
     /**
      * Set delivery status for outgoing group chat messages and files
      * 
-     * @param msgID Message ID
+     * @param msgId Message ID
      * @param contact The contact ID for which the entry is to be updated
      * @param status Status
      * @param reasonCode Reason code
